lesson6 1. в классе DateTimeHelper функция GetFormattedDate - работа с датой через кэш 2. поработал с Debug Color BlandedLayers - минимизировал наложение(смешение) цветов
