+++ VKClient/Cells/NewsCell/NewsLinkTableViewCell.xib
@@ -39,6 +39,9 @@
                         <fontDescription key="fontDescription" name="Charter-Roman" family="Charter" pointSize="14"/>
                         <color key="tintColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                         <size key="titleShadowOffset" width="2" height="2"/>
+                        <state key="normal">
+                            <color key="titleColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        </state>
                     </button>
                 </subviews>
                 <color key="backgroundColor" white="0.33333333329999998" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
